requirements.txt
pymilvus>=2.4.2
timm
torch
numpy
scikit-learn
pillow
streamlit

